+++ gpuRIR/extensions/filters/characteristic_filter.py
