+++ gpuRIR/extensions/hrtf/cipic_hrir/cipic_copyright.txt
